refactored my_for

+++ task/codewars/Productfib/pkg/productfib/productfib.go
@@ -16,11 +16,11 @@ type fib struct {
 // Else func returns 2 closest numbers, verifying F(n)*F(n+1) > prod and false
 func ProductFib(prod uint64) (res fib) {
 	for i := 1; ; i++ {
-		if sequences.Fibonacci(uint(i))*sequences.Fibonacci(uint(i+1)) == prod {
-			res.fib1, res.fib2, res.isequal = sequences.Fibonacci(uint(i)), sequences.Fibonacci(uint(i+1)), true
+		if sequences.Fibonacci(uint64(i))*sequences.Fibonacci(uint64(i+1)) == prod {
+			res.fib1, res.fib2, res.isequal = sequences.Fibonacci(uint64(i)), sequences.Fibonacci(uint64(i+1)), true
 			break
-		} else if sequences.Fibonacci(uint(i))*sequences.Fibonacci(uint(i+1)) > prod {
-			res.fib1, res.fib2, res.isequal = sequences.Fibonacci(uint(i)), sequences.Fibonacci(uint(i+1)), false
+		} else if sequences.Fibonacci(uint64(i))*sequences.Fibonacci(uint64(i+1)) > prod {
+			res.fib1, res.fib2, res.isequal = sequences.Fibonacci(uint64(i)), sequences.Fibonacci(uint64(i+1)), false
 			break
 		}
 	}
